Finished Project Push lol
I... I guess I finished the whole project before pushing once. I did this like a week ago too... Uhh it's a cool little pygame project that takes a triangle and turns it into an n-gon of matching size to the triangle. It also copies any drawings that you make to the triangle to the n-gon. Pretty much paper snowflakes.